+++ slang/lib/api/singleton.dart
@@ -2,26 +2,36 @@ import 'package:collection/collection.dart';
 import 'package:slang/api/locale.dart';
 import 'package:slang/api/pluralization.dart';
 import 'package:slang/api/state.dart';
+import 'package:slang/builder/builder/translation_model_builder.dart';
+import 'package:slang/builder/decoder/base_decoder.dart';
+import 'package:slang/builder/model/build_model_config.dart';
+import 'package:slang/builder/model/enums.dart';
+import 'package:slang/builder/utils/map_utils.dart';
+import 'package:slang/builder/utils/node_utils.dart';
 import 'package:slang/builder/utils/regex_utils.dart';
 
 /// Provides utility functions without any side effects.
-abstract class BaseAppLocaleUtils<E extends BaseAppLocale<T>,
-    T extends BaseTranslations> {
+abstract class BaseAppLocaleUtils<E extends BaseAppLocale<E, T>,
+    T extends BaseTranslations<E, T>> {
   /// Internal: The base locale
   final E baseLocale;
 
   /// Internal: All locales, unordered
   final List<E> locales;
 
+  /// Internal: Used for translation overrides
+  final BuildModelConfig? buildConfig;
+
   BaseAppLocaleUtils({
     required this.baseLocale,
     required this.locales,
+    this.buildConfig,
   });
 }
 
 // We use extension methods here to have a workaround for static members of the same name
-extension AppLocaleUtilsExt<E extends BaseAppLocale<T>,
-    T extends BaseTranslations> on BaseAppLocaleUtils<E, T> {
+extension AppLocaleUtilsExt<E extends BaseAppLocale<E, T>,
+    T extends BaseTranslations<E, T>> on BaseAppLocaleUtils<E, T> {
   /// Parses the raw locale to get the enum.
   /// Fallbacks to base locale.
   E parse(String rawLocale) {
@@ -84,10 +94,50 @@ extension AppLocaleUtilsExt<E extends BaseAppLocale<T>,
 
     return selected ?? baseLocale;
   }
+
+  /// Creates a translation instance with overrides stored in [content].
+  T buildWithOverrides({
+    required E locale,
+    required FileType fileType,
+    required String content,
+    PluralResolver? cardinalResolver,
+    PluralResolver? ordinalResolver,
+  }) {
+    return buildWithOverridesFromMap(
+      locale: locale,
+      map: BaseDecoder.getDecoderOfFileType(fileType).decode(content),
+      cardinalResolver: cardinalResolver,
+      ordinalResolver: ordinalResolver,
+    );
+  }
+
+  /// Creates a translation instance using the given [map].
+  T buildWithOverridesFromMap({
+    required E locale,
+    required Map map,
+    PluralResolver? cardinalResolver,
+    PluralResolver? ordinalResolver,
+  }) {
+    if (buildConfig == null) {
+      throw 'BuildConfig is null. Please generate the translations with <translation_overrides: true>';
+    }
+
+    final buildResult = TranslationModelBuilder.build(
+      buildConfig: buildConfig!,
+      localeDebug: locale.languageTag,
+      map: MapUtils.deepCast(map),
+    );
+
+    return locale.build(
+      overrides: buildResult.root.toFlatMap(),
+      cardinalResolver: cardinalResolver,
+      ordinalResolver: ordinalResolver,
+    );
+  }
 }
 
-abstract class BaseLocaleSettings<E extends BaseAppLocale<T>,
-    T extends BaseTranslations> {
+abstract class BaseLocaleSettings<E extends BaseAppLocale<E, T>,
+    T extends BaseTranslations<E, T>> {
   /// Locale enums sorted alphabetically and base locale first
   final List<E> locales;
 
@@ -106,11 +156,18 @@ abstract class BaseLocaleSettings<E extends BaseAppLocale<T>,
     required this.baseLocale,
     required this.utils,
   }) : this.translationMap = _buildMap(locales);
+
+  /// Updates the provider state and therefore triggers a rebuild
+  /// on all widgets listening to this provider.
+  ///
+  /// This is a flutter feature and this method will be overridden
+  /// by slang_flutter.
+  void updateProviderState(E locale, T translations) {}
 }
 
 // We use extension methods here to have a workaround for static members of the same name
-extension LocaleSettingsExt<E extends BaseAppLocale<T>,
-    T extends BaseTranslations> on BaseLocaleSettings<E, T> {
+extension LocaleSettingsExt<E extends BaseAppLocale<E, T>,
+    T extends BaseTranslations<E, T>> on BaseLocaleSettings<E, T> {
   /// Gets current locale.
   E get currentLocale {
     final locale = GlobalLocaleState.instance.getLocale();
@@ -144,6 +201,7 @@ extension LocaleSettingsExt<E extends BaseAppLocale<T>,
   /// This will be overwritten when using with flutter.
   E setLocale(E locale) {
     GlobalLocaleState.instance.setLocale(locale);
+    updateProviderState(locale, translationMap[locale]!);
     return locale;
   }
 
@@ -179,15 +237,52 @@ extension LocaleSettingsExt<E extends BaseAppLocale<T>,
     // update translation instances
     for (final curr in targetLocales) {
       translationMap[curr] = curr.build(
+        overrides: translationMap[curr]!.$meta.overrides, // keep old overrides
         cardinalResolver: cardinalResolver,
         ordinalResolver: ordinalResolver,
       );
     }
   }
+
+  /// Overrides existing translations of [locale] with new ones from [content].
+  /// Please do a try-catch to prevent app crashes!
+  void overrideTranslations({
+    required E locale,
+    required FileType fileType,
+    required String content,
+  }) {
+    final currentMetadata = translationMap[locale]!.$meta;
+    translationMap[locale] = utils.buildWithOverrides(
+      locale: locale,
+      content: content,
+      fileType: fileType,
+      cardinalResolver: currentMetadata.cardinalResolver,
+      ordinalResolver: currentMetadata.ordinalResolver,
+    );
+    if (locale == currentLocale) {
+      updateProviderState(locale, translationMap[locale]!);
+    }
+  }
+
+  /// Overrides existing translations of [locale] with new ones from the [map].
+  /// Please do a try-catch to prevent app crashes!
+  void overrideTranslationsFromMap({required E locale, required Map map}) {
+    final currentMetadata = translationMap[locale]!.$meta;
+    translationMap[locale] = utils.buildWithOverridesFromMap(
+      locale: locale,
+      map: map,
+      cardinalResolver: currentMetadata.cardinalResolver,
+      ordinalResolver: currentMetadata.ordinalResolver,
+    );
+    if (locale == currentLocale) {
+      updateProviderState(locale, translationMap[locale]!);
+    }
+  }
 }
 
-Map<E, T> _buildMap<E extends BaseAppLocale<T>, T extends BaseTranslations>(
-    List<E> locales) {
+Map<E, T>
+    _buildMap<E extends BaseAppLocale<E, T>, T extends BaseTranslations<E, T>>(
+        List<E> locales) {
   return <E, T>{
     for (final key in locales) key: key.build(),
   };