+++ slang/lib/builder/builder/generate_config_builder.dart
@@ -1,3 +1,4 @@
+import 'package:slang/builder/builder/build_model_config_builder.dart';
 import 'package:slang/builder/model/raw_config.dart';
 import 'package:slang/builder/model/generate_config.dart';
 import 'package:slang/builder/model/interface.dart';
@@ -5,24 +6,26 @@ import 'package:slang/builder/model/interface.dart';
 class GenerateConfigBuilder {
   static GenerateConfig build({
     required String baseName,
-    required RawConfig buildConfig,
+    required RawConfig config,
     required List<Interface> interfaces,
   }) {
     return GenerateConfig(
+      buildConfig: config.toBuildModelConfig(),
       baseName: baseName,
-      baseLocale: buildConfig.baseLocale,
-      fallbackStrategy: buildConfig.fallbackStrategy,
-      outputFormat: buildConfig.outputFormat,
-      localeHandling: buildConfig.localeHandling,
-      flutterIntegration: buildConfig.flutterIntegration,
-      translateVariable: buildConfig.translateVar,
-      enumName: buildConfig.enumName,
-      translationClassVisibility: buildConfig.translationClassVisibility,
-      renderFlatMap: buildConfig.renderFlatMap,
-      renderTimestamp: buildConfig.renderTimestamp,
-      contexts: buildConfig.contexts,
+      baseLocale: config.baseLocale,
+      fallbackStrategy: config.fallbackStrategy,
+      outputFormat: config.outputFormat,
+      localeHandling: config.localeHandling,
+      flutterIntegration: config.flutterIntegration,
+      translateVariable: config.translateVar,
+      enumName: config.enumName,
+      translationClassVisibility: config.translationClassVisibility,
+      renderFlatMap: config.renderFlatMap,
+      translationOverrides: config.translationOverrides,
+      renderTimestamp: config.renderTimestamp,
+      contexts: config.contexts,
       interface: interfaces,
-      imports: buildConfig.imports,
+      imports: config.imports,
     );
   }
 }