+++ slang/lib/builder/decoder/base_decoder.dart
@@ -4,8 +4,8 @@ import 'package:slang/builder/decoder/yaml_decoder.dart';
 import 'package:slang/builder/model/enums.dart';
 
 abstract class BaseDecoder {
-  /// Transforms the raw string (json, yaml, csv) to a standardized map structure
-  /// of Map<String, dynamic>
+  /// Transforms the raw string (json, yaml, csv)
+  /// to a standardized map structure of Map<String, dynamic>
   ///
   /// Children are Map<String, dynamic>, List<dynamic> or String
   ///