+++ slang/test/unit/builder/translation_model_builder_test.dart
@@ -13,7 +13,7 @@ void main() {
     test('1 StringTextNode', () {
       final result = TranslationModelBuilder.build(
         buildConfig: baseConfig.toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'test': 'a',
         },
@@ -29,7 +29,7 @@ void main() {
           keyCase: CaseStyle.snake,
           keyMapCase: CaseStyle.camel,
         ).toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'myMap': {'my_value': 'cool'},
         },
@@ -44,7 +44,7 @@ void main() {
           maps: ['my_map'],
           keyCase: CaseStyle.snake,
         ).toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'myMap': {'my_value 3': 'cool'},
         },
@@ -56,7 +56,7 @@ void main() {
     test('one link no parameters', () {
       final result = TranslationModelBuilder.build(
         buildConfig: baseConfig.toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'a': 'A',
           'b': 'Hello @:a',
@@ -70,7 +70,7 @@ void main() {
     test('one link 2 parameters straight', () {
       final result = TranslationModelBuilder.build(
         buildConfig: baseConfig.toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'a': r'A $p1 $p1 $p2',
           'b': 'Hello @:a',
@@ -84,7 +84,7 @@ void main() {
     test('linked translations with parameters recursive', () {
       final result = TranslationModelBuilder.build(
         buildConfig: baseConfig.toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'a': r'A $p1 $p1 $p2 @:b @:c',
           'b': r'Hello $p3 @:a',
@@ -100,7 +100,7 @@ void main() {
     test('linked translation with plural', () {
       final result = TranslationModelBuilder.build(
         buildConfig: baseConfig.toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'a': {
             'one': 'ONE',
@@ -126,7 +126,7 @@ void main() {
             generateEnum: true,
           ),
         ]).toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'a': {
             'male': 'MALE',
@@ -169,7 +169,7 @@ void main() {
             },
           ),
         ]).toBuildModelConfig(),
-        locale: defaultLocale,
+        localeDebug: defaultLocale.languageTag,
         map: {
           'myEntry': {
             'myList': [],